pyproject.toml
[tool.poetry]
name = "repl_python3_DiscordBotJokeTester"
version = "0.1.0"
description = ""
authors = ["RenatoJoao <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
python-decouple = "^3.4"
discord = "^1.0.1"
flask = "^1.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
